Added AutoPoke default config

+++ TS3Bot.AutoPoke/DTO/GroupDTO.cs
@@ -6,7 +6,7 @@ namespace TS3Bot.Ext.AutoPoke.DTO
 {
     public class GroupDTO
     {
-        public ushort Id { get; set; }
+        public uint Id { get; set; }
         public int DelayRelative { get; set; } = 0;
         public int DelayAbsolute { get; set; } = 0;
     }
